Start of spacebars compiler reorg
In advance of changing how {{> }} and {{# }} are compiled.

Goals for this reorg:
- "stache tags" are instanceof Spacebars.TemplateTag
- Phase out the word "stache"
- Better separate parsing, template tag expansion, and overall codegen and optimization using multiple files, and additional methods and properties on TemplateTag as appropriate

In particular, the things that happen to a template tag are it is parsed; validated (i.e. errors are thrown); and expanded (into code).  Validation and expansion can be methods on TemplateTag, as long as they also know the TEMPLATE_TAG_POSITION where the tag occurred.  They can be done either while parsing or while walking the AST for code generation, but if validation is done during parsing then it's easy to throw good error messages.  If expansion is done during code generation, then it's easier to inspect the parse tree for testing and debugging.

Also:
- Get rid of the word "Special" in html-tools (HTML.Special, getSpecialTag) in favor of TemplateTag.