Improve {{> foo}} lookup order
We now look for a helper named 'foo' (which can return
a dynamically chosen template object) before looking
for a global template named 'foo'. This is consistent
with the principle that adding a helper should generally
override other things with the same name (ie data properties)

This change was motivated by frontpage, which had both
a template and a helper named 'body'.