Remove unused cp_r preserveMode option
I had originally created a commit which let preserveMode reuse the
return value of the fs.statSync a few lines up in the file, but since
the code is dead, there was no way to test it. Removing the dead code
seemed like a better approach.  (If you want to revert this commit, I
recommend reusing the stat buffer!)