Reuse sha1 hashes calculated by the file watcher.
The file watcher (watch.js) calculates the sha1 hash of source files
to see if they've changed.  For static assets in `public`, the bundler
then calculates the same hash again for the client manifest.

This commit adds a small optimization to return the hash calculated by
the file watcher, so that it can be reused by the bundler.

There are more optimizations that probably could be done to avoid
other unnecessary sha1 hash recalculations, but they'd likely need
larger architectural changes.