Update update comment
Unless I'm confused, the code in the comment also runs in the pretty
typical case where you already have the latest release on your system
but the app you're in is pinned to an older release, and you run "meteor
update" without passing --release.  And that case was not described by
the previous comment.