make meteor-npm use runLog
this allows its messages to show up on the proxy, eg.

There's now a mix of runLog and buildmessage in this file, for two
reasons.

First, buildmessage's model is "build up a set of messages and then
display them all at once"; there's no streaming output, which is not
appropriate for the logUpdateDependencies call (which is mostly intended
to explain to users why their build is taking time).

Secondly, buildmessage doesn't yet have a great way of formatting big
blocks of text like "all the stderr from a failed npm invocation". If it
had such a feature, that would replace most of the uses of runLog in
this file.