A giant benchmarky test
It hangs right now because we didn't define any options for the constraint
solver which should be translated into a good cost function and a good
estimation function. That's why it just tries everything at the moment.

But this is a good test data to benchmark on to understand how different
optimizations and types of cost function, estimation function, or a combination
of two can affect the performance.