Remove `packageDirectoryForBuildInfo`.
We used to keep the `packageDirectoryForBuildInfo` of every package that
a package directly depends on in its buildinfo.json; if direct
dependencies change, then we might need to rebuild because they could
have registered or unregistered a plugin. Now we just keep the package
name and version (including build for our local packages) of our direct
dependencies.