Kill TBODY special case in DomRange
Background: When browsers parse HTML, they always insert a <tbody> around your <tr>s if you don't have one.  That is, `<table><tr>` becomes `<table><tbody><tr>`.  However, if you use the DOM API (appendChild et al.), this doesn't happen.  According to the latest specs and modern browsers, the TBODY element is not necessary in the DOM (though it is still always inserted when parsing HTML).  IE 6/7 disagrees, but that's history.  jQuery bends over backwards to insert TBODY elements dynamically, whenever you put a TR inside a TABLE.  Previously, we did too, but it's even harder for us.

This commit removes the code that special-cases tables in DomRange.  The code was never completely finished, and it will slow us down to have the extra essential complexity in DomRange.  Now, when you write `<table><tr>`, you'll get `<table><tr>` in the DOM tree.

It's important that we never optimize `<table><tr>...` into innerHTML, or else a template like `<table><tr><td>Foo</td></tr></table>` will get an inserted TBODY, while the same template with `{{foo}}` instead of `Foo` won't (because the table can't be optimized into static HTML).  A new test checks that we do this right.

The main downside of this change is that jQuery isn't as good for manipulating tables created by Meteor.  However, I think we can live with that.  jQuery made a particular choice there, and any problems that come up will be consistent across browsers and have workarounds (like inserting a <tbody> or not using a <table> tag).

Another downside concerns server-side rendering.  Rendering a template to a complete HTML string containing `<table><tr>` and then splatting that into the page will produce <tbody> tags that wouldn't be there if the page were rendered on the client.  We'll cross that bridge when we come to it.  In most ways (e.g. CSS rules), it shouldn't matter if there are <tbody> tags or not.