Allow circular deps when initializing catalog if already built.
Previously, we could only run any commands at all on this branch because
of a typo that was causing us to ignore all plugin dependencies when we
computed build-time dependencies (fixed in f94fd9d). This commit puts us
back in the state we thought we were in (things mostly work if you have
package builds cached from a different meteor branch).