Add a flag to MongoRunner which runs 3 mongods
This is only intended for testing. No data is preserved from one run to
the next, and if any mongod exits for any reason, the rest are killed;
there is no mongod restarting.

It takes a while (~20 seconds) to start up because it waits for the
replset to be ready.