Get plugins from implied dependencies
This means that standard-app-packages in your app is enough to get
'.html' and '.css' processing working, and apps work again.

This is a bit of a hack: we'd rather just make build time deps
directDependencies contain implied dependencies too instead of sticking
the entire constraint solver output in the database.