Fix flakiness in "defer in rendered callback" test
It was flaky before because template rendered callbacks get called
after flush time, but not if the template got destroyed in the meanwhile.

The way this test was written, if the client managed to respond to the server
rejecting the method before the client's flush cycle, the rendered callback
would never fire. Thus it would hang, since that callback was wrapped
in an expect.

Now we define a method on the client only, which makes it run as a stub
without the server rejecting the method (ever).