compare built tarball hash instead of buildid
buildids just change too damn much.  oh no, you changed a comment in
random which is transitively used by templating which processes .html, I
guess you need to bump all the versions of all packages with .html?
that sucks.  how about only changing it if the EFFECT of the build is
changed?

this does have some subtleties around platform-specific versions but we
think we have a good compromise