Revert "Add short aliases for '--add' and '--remove' options in "meteor authorized""
This reverts commit 32fc06e7f86b5f7b18226aa25d38d25af0461f9e.

As Glasser points out, our command option parsing is such that all `-a`
flags for commands have to be consistent in whether they are boolean or
not. Since we could easily imagine wanting a boolean `-a` or `-r` flag
in the future, we should be careful about adding these aliases. We'll
update the docs to remove `-a` and `-r` instead.