Fix audit-argument-checks error after login
If you successfully subscribed to a publisher, changed user ID, and the
second time running the publisher it failed the audit-argument-checks
test, the error would say "Did not check() all arguments during
publisher 'undefined'" instead of including the correct publisher name.

That's because the "recreate Subscription on setUserId" code forgot to
include the sub's name in the new Subscription (the name field is
literally only used for this error message).

Alluded to in #2092