Add _nestInCurrentComputation flag to UI.render.
Internally, Blaze sometimes relies on `UI.render`'s autoruns nesting
within the current computation, but it's often the case that callers
don't want `UI.render`'s autoruns to be stopped when the outer
computation is invalidated. This change fixes a bug in {{#each}}: if
`addedAt` ran inside a `Deps.autorun`, and then the outer computation
was invalidated, the corresponding item would stop updating because the
autoruns set up by `UI.render` would be stopped.

Fixes #2156.