Keep track of whether we are in <svg> as we traverse DOM.
Use this flag to determine whether we are creating elements inside an
<svg> element or not, instead of a list of SVG elements. The list of SVG
elements is problematic when we encounter a tag that is both an SVG
element and a normal HTML element; we'll end up creating it with
`document.createElementNS` instead of `document.createElement`.

See https://github.com/meteor/meteor/issues/1977.

This commit is untested; will come back to it after tests are passing on
blaze-refactor.