Fix a test of UI._templateInstance()
This actually morphs the test into a different test.  The old test wasn’t really valid; it checked that you could reach the rendered DOM of a helper from the templateInstance retrieved from the helper.  It only worked before because templateInstance.firstNode was an empty text node marker, so when the helper returned “first”, it would in effect save a pointer to the empty text node before the position where “first” would appear.