Fix Meteor._inherits to copy Parent static properties
Problem:
Static properties of the parent function ("class") were not copied to the child
function as expected. This is because the _.each underscore helper does not
propertly iterate over function properties.

Solution:
Don't use _.each to copy parent properties to the child function. Instead,
iterate directly and copy hasOwn properties to the child function.

An alternative solution is to make the _.each helper work properly with
functions in the underscore package.