Update comments after looking into e7250a0
It seems that when ObserveSequence observed a Cursor, it used to stop() the observe when the main autorun was invalidated, creating a “gap” during which no callbacks would be received (or fired).  This was before we used Deps.nonreactive in the main autorun (shielding the cursor.observe from being stopped).  Now the observe is only stopped upon re-run or when the ObserveSequence is stopped, and there is no gap.

Removed references to this gap from comments.

I believe the current code is correct, and in addition, we could now optimize the cursor-to-same-cursor case (and basically not do anything if seq===lastSeq and is a Cursor, i.e. not stop the old handle, create a new handle, or diff).