Extend options with encoding
Allow options object to have an encoding property that is used in the req_options. This allows for example options.encoding = null which will return the response content as buffer. In my case this allows me to use this package to save a generated pdf, from a post request to an external website, to file.