Make _onViewRendered private; onViewReady
The _onViewRendered hook fires when the View is actually rendered, not waiting until flush time or for the DOMRange to be attached.  It’s not very useful as part of the public API, because you can’t do anything with it (like manipulate the generated DOM).  Instead, expose onViewReady, which underlies the template “rendered” callback.